Revert "feat(outline): add outlineManager (#97)"
This reverts commit 0fd810d646e4eb2c271166c9ea5e34b01c3f6fe4.
